# Focus file before modifying DOM to avoid forced reflows