--- a/lib/archive-editor-view.js
+++ b/lib/archive-editor-view.js
@@ -68,7 +68,6 @@ export default class ArchiveEditorView {
         return
       }
 
-      this.refs.loadingMessage.style.display = 'none'
       if (error != null) {
         let message = 'Reading the archive file failed'
         if (error.message) {
@@ -80,6 +79,10 @@ export default class ArchiveEditorView {
         this.createTreeEntries(entries)
         this.updateSummary()
       }
+
+      // We hide the loading message _after_ creating the archive tree
+      // to avoid forced reflows.
+      this.refs.loadingMessage.style.display = 'none'
     })
   }
 
@@ -93,17 +96,22 @@ export default class ArchiveEditorView {
       if (entry.isDirectory()) {
         const entryView = new DirectoryView(this, index, this.path, entry)
         this.entries.push(entryView)
-        this.refs.tree.appendChild(entryView.element)
       } else {
         const entryView = new FileView(this, index, this.path, entry)
         this.entries.push(entryView)
-        this.refs.tree.appendChild(entryView.element)
       }
       index++
     }
 
-    this.refs.tree.style.display = ''
     this.selectFileAfterIndex(-1)
+
+    // Wait until selecting (focusing) the first file before appending the entries
+    // to avoid a double-forced reflow when focusing.
+    for (const entry of this.entries) {
+      this.refs.tree.appendChild(entry.element)
+    }
+
+    this.refs.tree.style.display = ''
   }
 
   updateSummary () {