--- a/lib/file-view.js
+++ b/lib/file-view.js
@@ -90,10 +90,16 @@ export default class FileView {
   }
 
   select () {
-    for (const selected of this.element.closest('.archive-editor').querySelectorAll('.selected')) {
-      selected.classList.remove('selected')
+    this.element.focus()
+
+    const archiveEditorElement = this.element.closest('.archive-editor')
+    // On initial tree creation, it is not possible for any entries to be selected
+    // (The entries also haven't been added to the DOM yet)
+    if (archiveEditorElement) {
+      for (const selected of archiveEditorElement.querySelectorAll('.selected')) {
+        selected.classList.remove('selected')
+      }
     }
     this.element.classList.add('selected')
-    this.element.focus()
   }
 }