--- a/styles/archive-view.less
+++ b/styles/archive-view.less
@@ -3,9 +3,10 @@
 .archive-editor {
   background-color: @inset-panel-background-color;
   overflow: auto;
+  contain: strict;
 
   .archive-container {
-    height:100%;
+    height: 100%;
     width: 100%;
 
     .inset-panel {