### Requirements

* Filling out the template is required. Any pull request that does not include enough information to be reviewed in a timely manner may be closed at the maintainers' discretion.
* All new code requires tests to ensure against regressions

### Description of the Change

Focusing an element is considered a "read" operation.  Since we were mutating the DOM before selecting the element, the focus call was causing a double forced reflow.  By waiting until after the focus call to mutate the DOM, we avoid those forced reflows, which were taking ~220ms combined.  While reflows still have to occur, they seem to be more efficient than the forced reflows as seen below.  I've also set `contain: strict` on the Archive View to improve layout performance.

| Before | After |
| ------- | ------ |
| 1484ms ![forced-reflows](https://user-images.githubusercontent.com/2766036/33776699-85479d74-dc42-11e7-94a6-92f09d1e7168.png) | 1200ms ![regular-reflow](https://user-images.githubusercontent.com/2766036/33780917-cf80ba78-dc52-11e7-90dc-6bddbdf1e834.png) |

### Alternate Designs

Don't focus a file on creation.  That however is not very user-friendly.

### Benefits

Removes a source of forced reflows and minorly improves the time it takes until the Archive View is usable.

### Possible Drawbacks

If an exception occurs while creating the tree entries or updating the summary, the loading view will not be removed.  Similarly with selecting the first file and actually adding the entries to the DOM.

### Applicable Issues

Refs #19