@50Wliu 2017-12-08
@simurai do you know anything about the `contain` property?  Reading MDN it seems like I can use `contain: strict` here and everything seems to work but I'd like to be sure that it won't horribly crash and burn in some edge cases.