@simurai 2017-12-08
I haven't experimented with `contain` yet. But since there is an `overflow: auto;` I *think* it's ok to use. There shouldn't be anything that can affect something outside of the overflow.

/cc @as-cii that has used `contain` for the editor.