@as-cii 2017-12-11
I haven't tested this but it would seem okay to use this approach. A couple of questions/comments:

* Forced reflows are not bad per se, unless they cause "layout thrashing" (e.g. read, write, read, write, etc.), which didn't seem to occur in the CPU profile you posted. Did the performance gain come from using stronger layout containment?
* I am unclear about the drawbacks of this approach and how they compare to the benefits this pull-request is providing. Could an exception occur while creating tree entries? How did we handle this situation before the proposed changes?

Thanks, @50Wliu!